pyproject.toml
[project]
name = "mood-journal-api"
version = "0.1.0"
description = "A FastAPI backend for a mood journal, using PostgreSQL."
authors = [
  {name = "404Skill", email = "your@email.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "SQLAlchemy==2.0.30",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
    "httpx==0.27.0"
]
